requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
openai
tiktoken
wikipedia
unstructured
